+++ pyriemann/datasets/sampling.py
@@ -250,12 +250,12 @@ def _slice_sampling(ptarget, n_samples, x0, n_burnin=20, thin=10,
     ptarget : function with one input
         The target pdf to sample from or a multiple of it.
     n_samples : int
-        How many samples to get from the ptarget distribution.
+        Number of samples to get from the ptarget distribution.
     x0 : ndarray
         Initial state for the MCMC procedure. Note that the shape of this array
         defines the dimensionality n_dim of the matrices to be sampled.
     n_burnin : int, default=20
-        How many samples to discard from the beginning of the chain generated
+        Number of samples to discard from the beginning of the chain generated
         by the slice sampling procedure. Usually the first samples are prone to
         non-stationary behavior and do not follow very well the target pdf.
     thin : int, default=10
@@ -268,7 +268,7 @@ def _slice_sampling(ptarget, n_samples, x0, n_burnin=20, thin=10,
         Pass an int for reproducible output across multiple function calls.
     n_jobs : int, default=1
         Number of jobs to use for the computation. This works by computing
-        each of the class centroid in parallel. If -1 all CPUs are used.
+        each sample in parallel. If -1 all CPUs are used.
 
     Returns
     -------
@@ -314,7 +314,7 @@ def _sample_parameter_r(n_samples, n_dim, sigma,
     Parameters
     ----------
     n_samples : int
-        How many samples to generate.
+        Number of samples to generate.
     n_dim : int
         Dimensionality of the SPD matrices to be sampled.
     sigma : float
@@ -323,7 +323,7 @@ def _sample_parameter_r(n_samples, n_dim, sigma,
         Pass an int for reproducible output across multiple function calls.
     n_jobs : int, default=1
         Number of jobs to use for the computation. This works by computing
-        each of the class centroid in parallel. If -1 all CPUs are used.
+        each sample in parallel. If -1 all CPUs are used.
     sampling_method : {"auto", "slice", "rejection"}, default="auto"
         Method used to sample parameter r: "auto", "slice" or "rejection".
         If "auto", sampling_method will be equal to "slice" for n_dim != 2 and
@@ -379,7 +379,7 @@ def _sample_parameter_U(n_samples, n_dim, random_state=None):
     Parameters
     ----------
     n_samples : int
-        How many samples to generate.
+        Number of samples to generate.
     n_dim : int
         Dimensionality of the SPD matrices to be sampled.
     random_state : int | RandomState instance | None, default=None
@@ -412,7 +412,7 @@ def _sample_gaussian_spd_centered(n_matrices, n_dim, sigma, random_state=None,
     Parameters
     ----------
     n_matrices : int
-        How many matrices to generate.
+        Number of matrices to generate.
     n_dim : int
         Dimensionality of the SPD matrices to be sampled.
     sigma : float
@@ -421,7 +421,7 @@ def _sample_gaussian_spd_centered(n_matrices, n_dim, sigma, random_state=None,
         Pass an int for reproducible output across multiple function calls.
     n_jobs : int, default=1
         Number of jobs to use for the computation. This works by computing
-        each of the class centroid in parallel. If -1 all CPUs are used.
+        each sample in parallel. If -1 all CPUs are used.
     sampling_method : {"auto", "slice", "rejection"}, default="auto"
         Method used to sample parameter r: "auto", "slice" or "rejection".
         If "auto", sampling_method will be equal to "slice" for n_dim != 2 and
@@ -490,7 +490,7 @@ def sample_gaussian_spd(n_matrices, mean, sigma, random_state=None,
     Parameters
     ----------
     n_matrices : int
-        How many matrices to generate.
+        Number of matrices to generate.
     mean : ndarray, shape (n_dim, n_dim)
         Center of the Riemannian Gaussian distribution.
     sigma : float | ndarray, shape (n_dim * (n_dim + 1) / 2, \
@@ -501,10 +501,12 @@ def sample_gaussian_spd(n_matrices, mean, sigma, random_state=None,
     random_state : int | RandomState instance | None, default=None
         Pass an int for reproducible output across multiple function calls.
     n_jobs : int, default=1
-        The number of jobs to use for the computation. This works by computing
-        each of the class centroid in parallel. If -1 all CPUs are used.
+        When sigma is a float,
+        the number of jobs to use for the computation. This works by computing
+        each sample in parallel. If -1 all CPUs are used.
     sampling_method : {"auto", "slice", "rejection"}, default="auto"
-        Method used to sample eigenvalues: "auto", "slice" or "rejection".
+        When sigma is a float,
+        method used to sample eigenvalues: "auto", "slice" or "rejection".
         If "auto", sampling_method will be equal to "slice" for n_dim != 2 and
         equal to "rejection" for n_dim = 2.
 
@@ -518,6 +520,8 @@ def sample_gaussian_spd(n_matrices, mean, sigma, random_state=None,
     Notes
     -----
     .. versionadded:: 0.3
+    .. versionchanged:: 0.11
+        Add support for dispersion defined as a covariance matrix.
 
     References
     ----------
@@ -526,7 +530,6 @@ def sample_gaussian_spd(n_matrices, mean, sigma, random_state=None,
         <https://hal.archives-ouvertes.fr/hal-01710191>`_
         S. Said, L. Bombrun, Y. Berthoumieu, and J. Manton. IEEE Trans Inf
         Theory, vol. 63, pp. 2153–2170, 2017.
-
     .. [2] `Wrapped gaussian on the manifold of symmetric positive
         definite matrices
         <https://openreview.net/pdf?id=EhStXG4dCS>`_
@@ -534,48 +537,46 @@ def sample_gaussian_spd(n_matrices, mean, sigma, random_state=None,
         Conference on Machine Learning (ICML), July 2025, Vancouver, Canada.
     """
 
-    n_dim = mean.shape[0]
+    n_dim, _ = mean.shape
 
-    if isinstance(sigma, np.ndarray):
-        # Sampling from the wrapped Gaussian distribution
-
-        if sigma.shape != (n_dim * (n_dim + 1) // 2,
-                           n_dim * (n_dim + 1) // 2):
-            raise ValueError("sigma must be a covariance matrix of shape  \
-                (n_dim * (n_dim + 1) / 2, n_dim * (n_dim + 1) / 2).")
-
-        rs = check_random_state(random_state)
-        # Sample from the multivariate normal distribution
-        t = rs.multivariate_normal(
-            size=n_matrices,
-            mean=np.zeros(n_dim * (n_dim + 1) // 2),
-            cov=sigma,
-        )
-
-        # Send to the tangent space at mean
-        mean_sqrt = sqrtm(mean)
-        samples_TS = mean_sqrt @ unupper(t) @ mean_sqrt
-
-        # Map back to the manifold
-        samples = exp_map_riemann(samples_TS, mean, Cm12=True)
-
-    elif isinstance(sigma, (int, float)):
-        # dispersion is corrected w.r.t. dimension
+    if isinstance(sigma, (int, float)):
+        # generate samples centered at identity
         samples_centered = _sample_gaussian_spd_centered(
             n_matrices=n_matrices,
             n_dim=n_dim,
-            sigma=sigma / np.sqrt(n_dim),
+            sigma=sigma / np.sqrt(n_dim),  # dispersion corrected w.r.t. dim
             random_state=random_state,
             n_jobs=n_jobs,
             sampling_method=sampling_method,
         )
 
-        # apply the parallel transport to mean on each of the samples
+        # apply the parallel transport from identity to mean on samples
         mean_sqrt = sqrtm(mean)
         samples = mean_sqrt @ samples_centered @ mean_sqrt
+
+    elif isinstance(sigma, np.ndarray):
+        n_ts = n_dim * (n_dim + 1) // 2
+        if sigma.shape != (n_ts, n_ts):
+            raise ValueError(
+                f"sigma must be a covariance matrix of shape ({n_ts}, {n_ts})."
+            )
+
+        # generate samples from the multivariate normal distribution
+        rs = check_random_state(random_state)
+        samples_ts_norm = rs.multivariate_normal(
+            size=n_matrices,
+            mean=np.zeros(n_ts),
+            cov=sigma,
+        )
+
+        # send the tangent space at mean
+        mean_sqrt = sqrtm(mean)
+        samples_ts = mean_sqrt @ unupper(samples_ts_norm) @ mean_sqrt
+        # map back to the manifold
+        samples = exp_map_riemann(samples_ts, mean, Cm12=True)
+
     else:
-        raise ValueError("sigma must be either a float or a covariance \
-                         matrix.")
+        raise ValueError("sigma must be either a float or a ndarray.")
 
     if not is_spsd(samples):
         msg = "Some of the sampled matrices are very badly conditioned and \